DOC: Adds polynomial convenience classes to web documenation
As a user, I was confused by the existence of both
`numpy.polynomial.polynomial.Polynomial` and
`numpy.polynomial.Polynomial`. I wasn't sure which to use.
The web docs contained no link to `numpy.polynomial.Polynomial,
so I assumed that I should use the longer name.
This PR adds these classes (no members, no docstrings) to the module page,
with a link to the usage instructions.

[skip actions] [skip azp] [skip cirrus]