DOD: Finished review of AI generated examples
Examples added to several functions in linalg. We can squash the
commits before pushing to numpy, and put a summary commit message.
Or we can push the entire commit train up so that the history shows.
Seeing what AI did, and what human revision happened, is visible
with the commit trail.

[skip actions] [skip azp] [skip cirrus]